+++ src/js/core/mejs.js
@@ -46,7 +46,7 @@ mejs.html5media = {
 	 */
 	mediaTypes: [
 		'audio/mp3', 'audio/ogg', 'audio/oga', 'audio/wav', 'audio/x-wav', 'audio/wave', 'audio/x-pn-wav', 'audio/mpeg', 'audio/mp4',
-		'video/mp4', 'video/webm', 'video/ogg'
+		'video/mp4', 'video/webm', 'video/ogg', 'video/ogv'
 	]
 };